refactor(types): centralize AttachmentMeta
Defines AttachmentMeta once in packages/db/src/schema/storage.ts (next to files) and re-exports from @pagespace/lib/types. Removes three duplicate inline shapes (channelMessages $type, attachment-utils.ts, channel messages route handler). Pure type refactor with no runtime change; sets up DM directMessages.attachmentMeta to import from the same canonical source without drifting.
